API-Tetsing/requirements.txt
langchain-aws>=0.1.0
boto3>=1.34.0
python-dotenv>=1.0.0
mcp-use>=0.1.0
streamlit>=1.28.0
python-dotenv>=1.0.0
asyncio
mcp_use
langchain>=0.1.0  
